requirements.txt
flask==2.3.3
requests==2.31.0
Werkzeug==2.3.7
gitpython==3.1.44